SharedData/requirements.txt
pydantic>=2.12.5
PyYAML>=6.0.3